Run digest auth test only against jvm
