+++ core/native/src/main/scala/sttp/client/MessageDigestCompatibility.scala
@@ -1,9 +1,6 @@
 package sttp.client
 
-import java.security.MessageDigest
-
 private[client] class MessageDigestCompatibility(algorithm: String) {
-  private lazy val md = MessageDigest.getInstance(algorithm)
-
-  def digest(input: Array[Byte]): Array[Byte] = md.digest(input)
+  def digest(input: Array[Byte]): Array[Byte] =
+    throw new UnsupportedOperationException("MessageDigest is currently not supported for native builds")
 }
