+++ core/shared/src/test/scala/sttp/client/testing/HttpTest.scala
@@ -252,15 +252,6 @@ trait HttpTest[F[_]]
         resp.body should be("Hello, adam!")
       }
     }
-
-    "perform digest authorization" in {
-      implicit val digestBackend =
-        new DigestAuthenticationBackend[F, Nothing, NothingT](backend, () => "e5d93287aa8532c1f5df9e052fda4c38")
-      val req = basicRequest.get(uri"$endpoint/secure_digest").response(asStringAlways).auth.digest("adam", "1234")
-      req.send()(digestBackend, implicitly[IsIdInRequest[Identity]]).toFuture().map { resp =>
-        resp.code shouldBe StatusCode.Ok
-      }
-    }
   }
 
   "compression" - {
